Extract grader library out of scripts/approach-b-framework/
The grader is consumed by both the approach-b-framework harness
(agent-maker.sh via grade-task.sh and test-spec-sh.sh) and the new
reverse-pipeline generator. Nesting it inside one consumer's
directory was misleading — it's a standalone library.

Move scripts/approach-b-framework/grader/ → scripts/grader/ and
update:
- grade-task.sh and test-spec-sh.sh to point at the new location
- reverse-pipeline imports (curation.ts, materialize.ts, test)
- pattern-spec.test.ts REPO_ROOT (4 '..' → 3)
- agent-maker.sh and AGENTS.md doc references

All 28 tests still pass; bash→dump-spec.ts smoke test confirms
end-to-end path resolution works from the new location.
