+++ scripts/approach-b-framework/agent-maker.sh
@@ -18,7 +18,7 @@
 #   agent-maker.sh --interactive -m google/gemini-3-flash-preview
 #
 # Task definition lives at scripts/approach-b-framework/tasks/<task>/test.yaml
-# (schema: grader/lib/test-spec.ts). The file declares which skill to
+# (schema: scripts/grader/lib/test-spec.ts). The file declares which skill to
 # invoke, the expected outcome (assembly + pattern, or gap), the
 # natural-language prompt, and the behavioral probe args.
 #
