+++ scripts/grader/__tests__/artifact.test.ts
