+++ scripts/grader/dump-spec.ts
