+++ scripts/grader/lib/artifact.ts
