+++ scripts/grader/lib/pattern-spec.ts
