+++ scripts/grader/lib/probes.ts
