+++ scripts/grader/lib/rubric.ts
