+++ scripts/grader/lib/types.ts
